feat(model): introduce SalaryEmployee with monthly earnings calculation (#12)
- Stores annualSalary (BigDecimal, >= 0)
- getEarnings(): annualSalary / 12 with scale(2), RoundingMode.HALF_UP
- Guards negative salary with IllegalArgumentException
- Adds SalaryEmployeeTest: happy path (50k→4166.67), zero, rounding, negative salary

+++ src/main/java/com/jaypatel/emanager/model/Employee.java
@@ -1,4 +1,7 @@
 package com.jaypatel.emanager.model;
+
+import java.math.BigDecimal;
+
 /**
  * Base abstract type for all employees in the system.
  * <p>
@@ -102,5 +105,5 @@ public abstract class Employee extends Person{
      *
      * @return the earnings amount
      */
-    public abstract double getEarnings();
+    public abstract BigDecimal getEarnings();
 }
